refactor: participant 패키지 이동
