+++ server/src/main/java/com/oddok/server/domain/participant/dao/ParticipantRepository.java
@@ -1,6 +1,6 @@
-package com.oddok.server.domain.bookmark.dao;
+package com.oddok.server.domain.participant.dao;
 
-import com.oddok.server.domain.studyroom.entity.Participant;
+import com.oddok.server.domain.participant.entity.Participant;
 import com.oddok.server.domain.user.entity.User;
 
 import java.util.Optional;
