+++ server/src/main/java/com/oddok/server/domain/studyroom/application/StudyRoomService.java
@@ -4,16 +4,15 @@ import com.oddok.server.common.errors.*;
 import com.oddok.server.domain.studyroom.dao.HashtagRepository;
 import com.oddok.server.domain.studyroom.dao.StudyRoomRepository;
 import com.oddok.server.domain.studyroom.dto.StudyRoomDto;
-import com.oddok.server.domain.bookmark.dao.ParticipantRepository;
+import com.oddok.server.domain.participant.dao.ParticipantRepository;
 import com.oddok.server.domain.studyroom.entity.Hashtag;
-import com.oddok.server.domain.studyroom.entity.Participant;
+import com.oddok.server.domain.participant.entity.Participant;
 import com.oddok.server.domain.studyroom.entity.StudyRoom;
 import com.oddok.server.domain.studyroom.mapper.StudyRoomMapper;
 import com.oddok.server.domain.user.dao.UserRepository;
 
 import com.oddok.server.domain.user.entity.User;
 
-import java.util.Objects;
 import java.util.Set;
 
 import lombok.RequiredArgsConstructor;
